+++ src/backend/resolvers/queries/events/event-meta-types.js
@@ -1,0 +1,4 @@
+module.exports = {
+  BASIC: 'basic',
+  RELATIONAL: 'relational',
+}