+++ src/frontend/components/EventLog/useEventLog.js
@@ -1,0 +1,26 @@
+import { useState, useEffect } from 'react'
+import { useQuery } from '@apollo/react-hooks'
+import _ from 'lodash'
+
+import { GET_EVENTS } from 'frontend/api/queries'
+
+import filterEvents from './filterEvents'
+
+const useEventLog = (filters) => {
+  const [events, setEvents] = useState([])
+  const { data, loading } = useQuery(GET_EVENTS)
+
+  useEffect(() => {
+    if (!_.isEmpty(data) && !loading) {
+      const events = filterEvents(data.events, filters)
+      setEvents(events)
+    }
+  }, [loading])
+
+  return {
+    events,
+    loading,
+  }
+}
+
+export default useEventLog